* server-src/cmdfile.c (cmdfile_write): Cast pid_t to int for printf.
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@5411 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/cmdfile.c
@@ -286,7 +286,7 @@ cmdfile_write(
 	char *holding_file = quote_string(cmddata->holding_file);
 	line = g_strdup_printf("%d FLUSH %s %s %s %s %s %s WORKING:%d %s\n",
 		id, config, holding_file, hostname, diskname,
-		dump_timestamp, storage_dest, cmddata->working_pid, status);
+		dump_timestamp, storage_dest, (int)cmddata->working_pid, status);
 	g_free(holding_file);
 	g_ptr_array_add(lines, line);
     } else if (cmddata->operation == CMD_COPY) {
@@ -295,7 +295,7 @@ cmdfile_write(
 	char *label = quote_string(cmddata->label);
 	line = g_strdup_printf("%d COPY %s %s %s %s %s %s %s %s WORKING:%d %s\n",
 		id, config, storage, pool, label, hostname, diskname,
-		dump_timestamp, storage_dest, cmddata->working_pid, status);
+		dump_timestamp, storage_dest, (int)cmddata->working_pid, status);
 	g_free(storage);
 	g_free(pool);
 	g_free(label);
